Changed the folder structure and started auth
