+++ DriverApp/package.json
@@ -10,8 +10,10 @@
   },
   "dependencies": {
     "@gluestack-ui/themed": "^0.1.15",
+    "@react-native-async-storage/async-storage": "^1.19.3",
     "@react-native-community/datetimepicker": "^7.5.0",
     "@react-navigation/stack": "^6.3.17",
+    "@reduxjs/toolkit": "^1.9.5",
     "axios": "^1.5.0",
     "dotenv": "^16.3.1",
     "expo": "^49.0.3",
@@ -33,7 +35,9 @@
     "react-native-select-dropdown": "^3.4.0",
     "react-native-svg": "^13.4.0",
     "react-native-table-component": "^1.2.2",
-    "react-native-web": "~0.19.6"
+    "react-native-web": "~0.19.6",
+    "react-redux": "^8.1.2",
+    "redux": "^4.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3",
