+++ DriverApp/src/components/Topic.jsx
@@ -1,6 +1,7 @@
 import styles from "./styles";
 import { View, Text } from "react-native";
 
+//component to add the topic for the needed page
 const Topic = ({ navigation, topic }) => {
   return (
     <View style={styles.Topic}>
